fix: add path formatting/sanitization for YAPI export
Ported formatPath() from legacy easy-yapi to ensure API paths sent to
YAPI start with '/' and only contain allowed characters.
Invalid characters are replaced with '/' and redundant slashes collapsed.
Also consolidated YapiFormatterExtendedTest into YapiFormatterTest.
